♻️ Refactor icons across the UI (#547, #554, #640)
