--- ui/src/icons/AdminEditPencilIcon.js
@@ -1,7 +1,0 @@
-import { BaseSvg } from 'icons';
-
-export default ({ fill = '#900000', ...props }) =>
-  BaseSvg({
-    svg: `<svg id="Layer_1" data-name="Layer 1" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 20.77 20.76"><defs><style>.cls-1{fill:${fill};}</style></defs><title>icon-pencil</title><path class="cls-1" d="M19.91,2.23,18.53.85a2.93,2.93,0,0,0-4.12,0L2,13.24a.64.64,0,0,0-.16.28L0,19.94a.62.62,0,0,0,.16.63.62.62,0,0,0,.46.19.55.55,0,0,0,.18,0L7.25,18.9a.66.66,0,0,0,.28-.16L19.91,6.36a2.91,2.91,0,0,0,0-4.13ZM6.73,17.7,1.59,19.17,3.06,14l10-10L16.7,7.73ZM19,5.44,17.62,6.82,14,3.15l1.37-1.38a1.64,1.64,0,0,1,2.3,0L19,3.15a1.58,1.58,0,0,1,.48,1.14A1.6,1.6,0,0,1,19,5.44Z"/></svg>`,
-    ...props,
-  });
