+++ ui/src/icons/EditIcon.js
